+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - new configurable setting: aling transform, valid options are: rigid, homography (#55)
 - full screen view menu option added to project window
 - full screen shortcut is Ctrl + Cmd + F or F11 on Windows (#57)
+- more shortcuts added for zoom in/out: +, =, Ctrl-+ and Ctrl-= could all zoom in, -, Ctrl-- could all zoom out (#57)
 
 ----